Fix 'Reject' button to reflect the journal coverage.
I fixed the `Reject` buttons for both the list and details views in
holdingpen to be displayed properly even if the workflow object
does not contain in `_extra_data` the required `journal_coverage` field.
This previously broke the button. Moreover, I introduced a small change
in the holdingpen angular controller that allowes buttons to be displayed
even if there is no `_extra_data`. Before this, the `ghasConflicts`
function did not allow the decision buttons to be displayed if `_extra_data`
is undefined.

Signed-off-by: Iuliana Voinea <iuliana.voinea@student.manchester.ac.uk>